Move CluterChangeListener to ReplicationEngine
