+++ ambry-replication/src/main/java/com/github/ambry/replication/ReplicationEngine.java
@@ -15,6 +15,7 @@ package com.github.ambry.replication;
 
 import com.codahale.metrics.MetricRegistry;
 import com.github.ambry.clustermap.ClusterMap;
+import com.github.ambry.clustermap.ClusterMapChangeListener;
 import com.github.ambry.clustermap.ClusterParticipant;
 import com.github.ambry.clustermap.DataNodeId;
 import com.github.ambry.clustermap.PartitionId;
@@ -24,6 +25,7 @@ import com.github.ambry.clustermap.ReplicaSyncUpManager;
 import com.github.ambry.commons.ResponseHandler;
 import com.github.ambry.config.ClusterMapConfig;
 import com.github.ambry.config.ReplicationConfig;
+import com.github.ambry.config.StoreConfig;
 import com.github.ambry.network.ConnectionPool;
 import com.github.ambry.notification.NotificationSystem;
 import com.github.ambry.server.StoreManager;
@@ -47,11 +49,14 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.locks.ReadWriteLock;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
 import java.util.function.Predicate;
+import java.util.stream.Collectors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -66,8 +71,10 @@ import static com.github.ambry.clustermap.VcrClusterParticipant.*;
  * 4. Initialize and shutdown all the components required to perform replication
  */
 public abstract class ReplicationEngine implements ReplicationAPI {
-
+  protected CountDownLatch startupLatch = new CountDownLatch(1);
+  protected boolean started = false;
   protected final ReplicationConfig replicationConfig;
+  protected final StoreConfig storeConfig;
   protected final ClusterMap clusterMap;
   protected final ScheduledExecutorService scheduler;
   private final AtomicInteger correlationIdGenerator;
@@ -102,23 +109,26 @@ public abstract class ReplicationEngine implements ReplicationAPI {
   protected LeaderBasedReplicationAdmin leaderBasedReplicationAdmin = null;
 
   public ReplicationEngine(ReplicationConfig replicationConfig, ClusterMapConfig clusterMapConfig,
-      StoreKeyFactory storeKeyFactory, ClusterMap clusterMap, ScheduledExecutorService scheduler, DataNodeId dataNode,
-      List<? extends ReplicaId> replicaIds, ConnectionPool connectionPool, MetricRegistry metricRegistry,
-      NotificationSystem requestNotification, StoreKeyConverterFactory storeKeyConverterFactory,
-      String transformerClassName, ClusterParticipant clusterParticipant, StoreManager storeManager,
-      Predicate<MessageInfo> skipPredicate) throws ReplicationException {
-    this(replicationConfig, clusterMapConfig, storeKeyFactory, clusterMap, scheduler, dataNode, replicaIds,
+      StoreConfig storeConfig, StoreKeyFactory storeKeyFactory, ClusterMap clusterMap,
+      ScheduledExecutorService scheduler, DataNodeId dataNode, List<? extends ReplicaId> replicaIds,
+      ConnectionPool connectionPool, MetricRegistry metricRegistry, NotificationSystem requestNotification,
+      StoreKeyConverterFactory storeKeyConverterFactory, String transformerClassName,
+      ClusterParticipant clusterParticipant, StoreManager storeManager, Predicate<MessageInfo> skipPredicate)
+      throws ReplicationException {
+    this(replicationConfig, clusterMapConfig, storeConfig, storeKeyFactory, clusterMap, scheduler, dataNode, replicaIds,
         connectionPool, metricRegistry, requestNotification, storeKeyConverterFactory, transformerClassName,
         clusterParticipant, storeManager, skipPredicate, null, SystemTime.getInstance());
   }
 
   public ReplicationEngine(ReplicationConfig replicationConfig, ClusterMapConfig clusterMapConfig,
-      StoreKeyFactory storeKeyFactory, ClusterMap clusterMap, ScheduledExecutorService scheduler, DataNodeId dataNode,
-      List<? extends ReplicaId> replicaIds, ConnectionPool connectionPool, MetricRegistry metricRegistry,
-      NotificationSystem requestNotification, StoreKeyConverterFactory storeKeyConverterFactory,
-      String transformerClassName, ClusterParticipant clusterParticipant, StoreManager storeManager,
-      Predicate<MessageInfo> skipPredicate, FindTokenHelper findTokenHelper, Time time) throws ReplicationException {
+      StoreConfig storeConfig, StoreKeyFactory storeKeyFactory, ClusterMap clusterMap,
+      ScheduledExecutorService scheduler, DataNodeId dataNode, List<? extends ReplicaId> replicaIds,
+      ConnectionPool connectionPool, MetricRegistry metricRegistry, NotificationSystem requestNotification,
+      StoreKeyConverterFactory storeKeyConverterFactory, String transformerClassName,
+      ClusterParticipant clusterParticipant, StoreManager storeManager, Predicate<MessageInfo> skipPredicate,
+      FindTokenHelper findTokenHelper, Time time) throws ReplicationException {
     this.replicationConfig = replicationConfig;
+    this.storeConfig = storeConfig;
     this.storeKeyFactory = storeKeyFactory;
     if (findTokenHelper == null) {
       try {
@@ -150,6 +160,8 @@ public abstract class ReplicationEngine implements ReplicationAPI {
     this.skipPredicate = skipPredicate;
     replicaSyncUpManager = clusterParticipant == null ? null : clusterParticipant.getReplicaSyncUpManager();
     this.time = time;
+
+    clusterMap.registerClusterMapListener(new ClusterMapChangeListenerImpl());
   }
 
   /**
@@ -376,8 +388,8 @@ public abstract class ReplicationEngine implements ReplicationAPI {
    * @return The name of the thread.
    */
   protected String getReplicaThreadName(String datacenterToReplicateFrom, int threadIndexWithinPool) {
-    return "ReplicaThread-" + (dataNodeId.getDatacenterName().equals(datacenterToReplicateFrom) ? "Intra-"
-            : "Inter-") + threadIndexWithinPool + "-" + datacenterToReplicateFrom;
+    return "ReplicaThread-" + (dataNodeId.getDatacenterName().equals(datacenterToReplicateFrom) ? "Intra-" : "Inter-")
+        + threadIndexWithinPool + "-" + datacenterToReplicateFrom;
   }
 
   /**
@@ -670,6 +682,115 @@ public abstract class ReplicationEngine implements ReplicationAPI {
           new HashSet<>(leaderPartition.getReplicaIdsByState(ReplicaState.LEADER, null));
       peerLeaderReplicaSet.remove(localReplica);
       return peerLeaderReplicaSet;
+    }
+  }
+
+  /**
+   * Implementation of {@link ClusterMapChangeListener} that helps replication manager react to cluster map changes.
+   */
+  class ClusterMapChangeListenerImpl implements ClusterMapChangeListener {
+
+    /**
+     * {@inheritDoc}
+     * Note that, this method should be thread-safe because multiple threads (from different cluster change handlers) may
+     * concurrently update remote replica infos.
+     */
+    @Override
+    public void onReplicaAddedOrRemoved(List<ReplicaId> addedReplicas, List<ReplicaId> removedReplicas) {
+      // 1. wait for start() to complete
+      try {
+        startupLatch.await();
+      } catch (InterruptedException e) {
+        logger.warn("Waiting for startup is interrupted.");
+        throw new IllegalStateException("Replication manager startup is interrupted while updating remote replicas");
+      }
+      if (started) {
+        // Read-write lock avoids contention between addReplica()/removeReplica() and onReplicaAddedOrRemoved() methods.
+        // Read lock for current method should suffice because multiple threads from cluster change handlers should be able
+        // to access partitionToPartitionInfo map. Each thead only updates PartitionInfo of certain partition and synchronization
+        // is only required within PartitionInfo. Also, addRemoteReplicaInfoToReplicaThread() is thread-safe which allows
+        // several threads from cluster change handlers to add remoteReplicaInfo
+        rwLock.readLock().lock();
+        try {
+          // 2. determine if added/removed replicas have peer replica on local node.
+          //    We skip the replica on current node because it should already be added/removed by state transition thread.
+          Set<ReplicaId> addedPeerReplicas = addedReplicas.stream()
+              .filter(r -> partitionToPartitionInfo.containsKey(r.getPartitionId()) && r.getDataNodeId() != dataNodeId)
+              .collect(Collectors.toSet());
+          Set<ReplicaId> removedPeerReplicas = removedReplicas.stream()
+              .filter(r -> partitionToPartitionInfo.containsKey(r.getPartitionId()) && r.getDataNodeId() != dataNodeId)
+              .collect(Collectors.toSet());
+
+          // No additional synchronization is required because cluster change handler of each dc only updates replica-threads
+          // belonging to certain dc. Hence, there is only one thread adding/removing remote replicas within a certain dc.
+
+          // 3. create replicaInfo for new remote replicas and assign them to replica-threads.
+          List<RemoteReplicaInfo> replicaInfosToAdd = new ArrayList<>();
+          for (ReplicaId remoteReplica : addedPeerReplicas) {
+            logger.info("Attempting to add remote replica in replication manager: " + remoteReplica);
+            PartitionInfo partitionInfo = partitionToPartitionInfo.get(remoteReplica.getPartitionId());
+            // create findToken, remoteReplicaInfo
+            FindToken findToken =
+                tokenHelper.getFindTokenFactoryFromReplicaType(remoteReplica.getReplicaType()).getNewFindToken();
+            RemoteReplicaInfo remoteReplicaInfo =
+                new RemoteReplicaInfo(remoteReplica, partitionInfo.getLocalReplicaId(), partitionInfo.getStore(),
+                    findToken,
+                    TimeUnit.SECONDS.toMillis(storeConfig.storeDataFlushIntervalSeconds) * Replication_Delay_Multiplier,
+                    SystemTime.getInstance(), remoteReplica.getDataNodeId().getPortToConnectTo());
+            logger.info("Adding remote replica {} on {} to partition info.", remoteReplica.getReplicaPath(),
+                remoteReplica.getDataNodeId());
+            if (partitionInfo.addReplicaInfoIfAbsent(remoteReplicaInfo)) {
+              replicationMetrics.addMetricsForRemoteReplicaInfo(remoteReplicaInfo,
+                  replicationConfig.replicationTrackPerDatacenterLagFromLocal);
+              replicaInfosToAdd.add(remoteReplicaInfo);
+            }
+          }
+          addRemoteReplicaInfoToReplicaThread(replicaInfosToAdd, true);
+
+          // 4. remove replicaInfo from existing partitionInfo and replica-threads
+          List<RemoteReplicaInfo> replicaInfosToRemove = new ArrayList<>();
+          for (ReplicaId remoteReplica : removedPeerReplicas) {
+            logger.info("Attempting to remove remote replica from replication manager: " + remoteReplica);
+            PartitionInfo partitionInfo = partitionToPartitionInfo.get(remoteReplica.getPartitionId());
+            RemoteReplicaInfo removedReplicaInfo = partitionInfo.removeReplicaInfoIfPresent(remoteReplica);
+            if (removedReplicaInfo != null) {
+              replicationMetrics.removeMetricsForRemoteReplicaInfo(removedReplicaInfo);
+              logger.info("Removing remote replica {} on {} from replica threads.", remoteReplica.getReplicaPath(),
+                  remoteReplica.getDataNodeId());
+              replicaInfosToRemove.add(removedReplicaInfo);
+            }
+          }
+          removeRemoteReplicaInfoFromReplicaThread(replicaInfosToRemove);
+          logger.info("{} peer replicas are added and {} peer replicas are removed in replication manager",
+              addedPeerReplicas.size(), removedPeerReplicas.size());
+        } finally {
+          rwLock.readLock().unlock();
+        }
+      }
+    }
+
+    /**
+     * {@inheritDoc}
+     * Note that, this method should be thread-safe because multiple threads (from different cluster change handlers) may
+     * concurrently call this method.
+     */
+    @Override
+    public void onRoutingTableChange() {
+
+      // wait for start() to complete
+      try {
+        startupLatch.await();
+      } catch (InterruptedException e) {
+        logger.warn("Waiting for startup is interrupted.");
+        throw new IllegalStateException(
+            "Replication manager startup is interrupted while handling routing table change");
+      }
+
+      if (leaderBasedReplicationAdmin != null) {
+        // Refreshes the remote leader information for all local leader partitions maintained in the in-mem structure in LeaderBasedReplicationAdmin.
+        // Thread safety is ensured in LeaderBasedReplicationAdmin::refreshPeerLeadersForAllPartitions().
+        leaderBasedReplicationAdmin.refreshPeerLeadersForLeaderPartitions();
+      }
     }
   }
 }
